Fix/#26: 선호 옵션 수정
기존의 ContentBox 컴포넌트를 분리
 - PreferContent 로 분리
